Create functionality for adding objects to the user.
Made the 'Create' button pick up on the type of thing being created by looking at the title of the form... idk if this is okay or not.
Objects get created and are added to the user, and then the form is closed.

Also fixed a styling error I made --> the textarea elements had grey text instead of the --border colour they should have had.